타 빌딩 데이터 생성 값 조절

+++ src/main/java/com/ThreeZem/three_zem_back/service/AnalysisService.java
@@ -182,7 +182,7 @@ public class AnalysisService {
             long ourYearlyPrice = 0;
             for (EnergyReadingDto energyReadingDto : ourPrice) {
                 if (energyReadingDto != null && energyReadingDto.getDatas() != null && !energyReadingDto.getDatas().isEmpty()) {
-                    ourYearlyPrice += energyReadingDto.getDatas().get(0).getUsage();
+                    ourYearlyPrice += (long) energyReadingDto.getDatas().get(0).getUsage();
                 }
             }
             billAnalysisDto.setOurYearyPrice(ourYearlyPrice);
